refactor: move vite workaround into `nitro:generate` hook
