+++ packages/nuxt3/src/core/nitro.ts
@@ -20,6 +20,7 @@ export function initNitro (nuxt: Nuxt) {
   nuxt.hooks.addHooks(nitroContext.nuxtHooks)
   nuxt.hook('close', () => nitroContext._internal.hooks.callHook('close'))
   nitroContext._internal.hooks.hook('nitro:document', template => nuxt.callHook('nitro:document', template))
+  nitroContext._internal.hooks.hook('nitro:generate', ctx => nuxt.callHook('nitro:generate', ctx))
 
   // @ts-ignore
   nuxt.hooks.addHooks(nitroDevContext.nuxtHooks)
